refactor: share form field utils and update search form
