+++ client/src/components/admin/AirportManagement.js
@@ -12,7 +12,8 @@ import {
 	deleteAllAirports,
 } from '../../redux/actions/airport';
 import { fetchCountries } from '../../redux/actions/country';
-import { FIELD_TYPES, createAdminManager } from './utils';
+import { createAdminManager } from './utils';
+import { FIELD_TYPES } from '../utils';
 import { FIELD_LABELS, UI_LABELS, VALIDATION_MESSAGES } from '../../constants';
 
 const AirportManagement = () => {
