+++ client/src/components/admin/CountryManagement.js
@@ -10,7 +10,8 @@ import {
 	deleteCountry,
 	deleteAllCountries,
 } from '../../redux/actions/country';
-import { FIELD_TYPES, createAdminManager } from './utils';
+import { createAdminManager } from './utils';
+import { FIELD_TYPES } from '../utils';
 import { FIELD_LABELS, UI_LABELS, VALIDATION_MESSAGES } from '../../constants';
 
 const CountryManagement = () => {
