+++ client/src/components/admin/PassengerManagement.js
@@ -19,7 +19,8 @@ import {
 } from '../../redux/actions/bookingPassenger';
 import { fetchBookings } from '../../redux/actions/booking';
 import { fetchCountries } from '../../redux/actions/country';
-import { FIELD_TYPES, createAdminManager } from './utils';
+import { createAdminManager } from './utils';
+import { FIELD_TYPES } from '../utils';
 import { formatDate, validateDate, isDuplicateInBooking } from '../utils';
 import { ENUM_LABELS, FIELD_LABELS, UI_LABELS, VALIDATION_MESSAGES, getEnumOptions } from '../../constants';
